+++ CHANGELOG.markdown
@@ -1,9 +1,10 @@
 4.13
 ----
 * Pattern synonyms
-* Moved `foldMapBy` and `foldBy` into `reflection`.
+* Moved `foldMapBy` and `foldBy` into `reflection` 2.1
 * Added `traverseByOf`.
-* Modified the signatures of `alaf` and `auf` to work with a `Functor` rather than a `Profunctor`.
+* Modified the signatures of `alaf` and `auf` to work with a `Functor` rather than a `Profunctor` and rather drastically generalized them.
+* Removed `Control.Lens.Internal.Getter.coerce` in favor of the upstream `phantom` combinator in `contravariant` 1.3+
 
 4.12.3
 ------